# .openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CWS Next App API",
    "version": "0.1.0",
    "description": "Automated API documentation for the CWS Next App. This document is generated from Zod schemas and route handler metadata.",
    "contact": {
      "name": "API Support",
      "email": "api@crossweavesourcing.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
...
  "paths": {
...
    "/api/auth/webauthn/login-verify": {
      "post": {
        "operationId": "webauthnLoginVerify",
        "summary": "Verify WebAuthn login response",
        "description": "Verifies a passwordless passkey response for this enrolled device. On success, issues dashboard session cookies or redirects to email verification for high-risk sign-ins.",
        "tags": [
          "Authentication"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$",
                    "description": "Account email used to start passkey login",
                    "example": "admin@example.com"
                  },
                  "response": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "description": "Authenticator credential ID",
                        "example": "dBjftJeZ4CVP-mB92K27uhbUJU1p1r_wW1gFWFOEjXk"
                      },
                      "rawId": {
                        "type": "string",
                        "description": "Raw credential ID (base64url)"
                      },
                      "response": {
                        "type": "object",
                        "properties": {
                          "authenticatorData": {
                            "type": "string",
                            "description": "Authenticator data (base64url)"
                          },
                          "clientDataJSON": {
                            "type": "string",
                            "description": "Client data JSON (base64url)"
                          },
                          "signature": {
                            "type": "string",
                            "description": "Signature (base64url)"
                          },
                          "userHandle": {
                            "description": "User handle (base64url)",
                            "type": "string"
                          }
                        },
                        "required": [
                          "authenticatorData",
                          "clientDataJSON",
                          "signature"
                        ],
                        "additionalProperties": false
                      },
                      "type": {
                        "type": "string",
                        "const": "public-key",
                        "example": "public-key"
                      },
                      "clientExtensionResults": {
                        "type": "object",
                        "propertyNames": {
                          "type": "string"
                        },
                        "additionalProperties": {}
                      }
                    },
                    "required": [
                      "id",
                      "rawId",
                      "response",
                      "type"
                    ],
                    "additionalProperties": false
                  }
                },
                "required": [
                  "email",
                  "response"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful — session cookies set",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "redirect": {
                      "type": "string",
                      "example": "/dashboard/verify-2fa?method=email"
                    }
                  },
                  "required": [
                    "success"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid WebAuthn response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Passkey is not available for this device or challenge expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/login": {
      "post": {
        "operationId": "webPasswordLogin",
        "summary": "Authenticate the dashboard password form",
        "description": "Same-origin HTML form endpoint. On completion it returns a 303 redirect to the dashboard, an MFA/password-change step, or the login page with a safe error code.",
        "tags": [
          "Authentication"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  },
                  "rememberMe": {
                    "type": "string",
                    "enum": [
                      "on",
                      "true"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "303": {
            "description": "Authentication flow redirect; secure cookies may be set."
          }
        }
      }
    },
...
    "/api/mobile/v1/auth/mfa/webauthn/verify": {
      "post": {
        "operationId": "mobileMfaWebauthnVerify",
        "summary": "Deprecated WebAuthn MFA verification",
        "description": "Deprecated. Passkeys are available for passwordless sign-in only; use email code for verification.",
        "deprecated": true,
        "tags": [
          "Mobile Authentication"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "challengeToken": {
                    "type": "string",
                    "minLength": 1,
                    "description": "Challenge token from login response",
                    "example": "eyJhbGciOiJIUzI1NiJ9..."
                  },
                  "response": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "description": "Authenticator credential ID"
                      },
                      "rawId": {
                        "type": "string",
                        "description": "Raw credential ID (base64url)"
                      },
                      "response": {
                        "type": "object",
                        "properties": {
                          "authenticatorData": {
                            "type": "string",
                            "description": "Authenticator data (base64url)"
                          },
                          "clientDataJSON": {
                            "type": "string",
                            "description": "Client data JSON (base64url)"
                          },
                          "signature": {
                            "type": "string",
                            "description": "Signature (base64url)"
                          },
                          "userHandle": {
                            "description": "User handle (base64url)",
                            "type": "string"
                          }
                        },
                        "required": [
                          "authenticatorData",
                          "clientDataJSON",
                          "signature"
                        ],
                        "additionalProperties": false
                      },
                      "type": {
                        "type": "string",
                        "const": "public-key",
                        "example": "public-key"
                      },
                      "clientExtensionResults": {
                        "type": "object",
                        "propertyNames": {
                          "type": "string"
                        },
                        "additionalProperties": {}
                      }
                    },
                    "required": [
                      "id",
                      "rawId",
                      "response",
                      "type"
                    ],
                    "additionalProperties": false
                  }
                },
                "required": [
                  "challengeToken",
                  "response"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "410": {
            "description": "WebAuthn MFA is no longer supported",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Content-Type must be application/json",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/mobile/v1/auth/passkeys/register/verify": {
      "post": {
        "operationId": "mobilePasskeyRegisterVerify",
        "summary": "Deprecated mobile passkey registration verification",
        "description": "Mobile passkey registration is unavailable until mobile device binding is supported.",
        "deprecated": true,
        "tags": [
          "Mobile Authentication"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid registration response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized or expired challenge",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "410": {
            "description": "Mobile passkeys unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/mobile/v1/auth/passkeys/login/verify": {
      "post": {
        "operationId": "mobilePasskeyLoginVerify",
        "summary": "Deprecated mobile passkey login verification",
        "description": "Mobile passkey login is unavailable until mobile device binding is supported.",
        "deprecated": true,
        "tags": [
          "Mobile Authentication"
        ],
        "security": [],
        "responses": {
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired challenge",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "410": {
            "description": "Mobile passkeys unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "cws_session",
        "description": "HMAC-signed session cookie issued by the web dashboard login flow. Obtained via POST /api/auth/login (Server Action) or Google OAuth callback."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Ed25519-signed JWT issued by the mobile authentication endpoints. Obtained via POST /api/mobile/v1/auth/password or /api/mobile/v1/auth/google."
      },
      "pendingSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "cws_2fa_pending",
        "description": "Short-lived HMAC-signed pending session cookie set during MFA challenge. Valid for 5 minutes. Used by WebAuthn and TOTP verification endpoints."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "A human-readable error message",
            "example": "Resource not found"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
    }
  }
}